Decompose InvoiceWorkflowService into smaller services
