+++ backend/Glovelly.Api/Services/IInvoiceNumberService.cs
@@ -1,0 +1,6 @@
+namespace Glovelly.Api.Services;
+
+public interface IInvoiceNumberService
+{
+    Task<string> GenerateInvoiceNumberAsync(DateOnly invoiceDate, CancellationToken cancellationToken = default);
+}
